Step 1: Register Block
Here we hook into `acf/init` to ensure Advanced Custom Fields has been initialized
Then we register the block using the ACF function `acf_register_block`
This registers a block to Gutenberg, but also adds the block as a new Context for ACF Fields to be applied to.
The `render_callback` is a standard Gutenberg block attribute that allows for dynamic rendering of blocks. Here, we define our callback to be `acf_testimonial_callback`, and inside that function we output some simple markup.

At this point, we can now add the `ACF Testimonial Block` in Gutenberg, and we will see the name of the block rendered in the Editor and in our Theme when the post is saved.