Step 2: Block Markup and Styles
In this step, we write the markup we want to use to output our Block in our theme. This will also be the same markup used for previewing the component while editing in Gutenberg.

We define a few variables that we will want to be editable when interacting with the block: $testimonial, $alignment, $avatar_url, $name, $text_color and $background_color. For now, we've hard-coded their values so we can test our markup and make sure it looks ok.

For styling, we simply output a <style> tag along with our block markup, and we use some of our variables to make the styles dynamic.

At this point, we can see our Block taking shape, both in the Gutenberg UI and in our Theme when we save a post and look at it.