Step 3: Register Fields
In this step we register our fields. You can use the Advanced Custom Fields UI to register your fields, or you can register the fields via code, like we've done here. . .There's also an option to use JSON for the field registration.

Here we've registered the following fields:
- testimonial
- avatar
- name
- background_color
- text_color
- alignment

Then, in our `acf_testimonial_callback` function, we use the values of the fields to replace our hard-coded variables.

Now, we have a fully interactive block with fields we can edit interactively and we see live previews of our changes, and our theme renders the dynamic values in our markup and styles.

Neat!